refactoring base self test class into utils
