+++ dot/self_tests/unit/test_connection_utils.py
@@ -1,4 +1,4 @@
-from .base_self_test_class import BaseSelfTestClass
+from ..utils.base_self_test_class import BaseSelfTestClass
 
 # UT after base_self_test_class imports
 from utils.connection_utils import (  # pylint: disable=wrong-import-order
